First time I used the XML-Editor, which does not work properly for me sometimes :-(
changed BR-description of SA vz.58, because original description could be misunderstood as being offensive.
changed names of G36E and G36KE to G36 Tactical and G36K Tactical, because they no longer have the carrying handle with a small sight, they have a Picatinny Rail instead.
Removed built-in Reflex Scope/Sight on SIG SG551 and SG552. These guns are now able to mount other scopes as well. The pics did not show the scope/sight, so it was not implied that they are there.

git-svn-id: https://ja2svn.mooo.com/source/ja2_v1.13_data@718 4f8fa57e-7814-0410-bad4-adc449f26b7c