New Ini for item property modifiers
The new Item_Settings.ini is there to host modifiers for item properties in the different XML files. For now it is only used for the Weapon Overheating feature to adjust parameters based on gun class.
The ini can hold any modifiers, for example modifiers for range, armor rating, damage, weight, AP to load/fire etc. Of course somebody has to add that to the game code first. ;)


git-svn-id: https://ja2svn.mooo.com/source/ja2_v1.13_data@1800 4f8fa57e-7814-0410-bad4-adc449f26b7c